file module now building and has 4 functions. file_test.go has been made, the functions in builtin_fn_io.go relating to files have been commented out and marked depreacted.
